Minor: Extract ExecutionPlanVisitor to its onw module
